search.elixpo/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.1
attrs==25.4.0
av==16.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.3
brotli==1.2.0
build==1.4.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.3.0
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.3
ctranslate2==4.6.3
cycler==0.12.1
cymem==2.0.13
ddgs==9.8.0
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
fastapi==0.121.2
faster-whisper==1.2.1
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
geographiclib==2.1
geopy==2.4.1
git-filter-repo==2.47.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
h3==4.3.1
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
kubernetes==35.0.0
limits==5.6.0
llvmlite==0.45.1
loguru==0.7.3
lxml==6.0.2
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.15
networkx==3.5
nltk==3.9.2
nodejs-wheel-binaries==24.11.1
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
openai-whisper==20250625
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==25.0
pillow==12.0.0
playwright==1.55.0
posthog==7.8.6
preshed==3.0.12
primp==0.15.0
priority==2.0.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytube==15.0.0
pytubefix==10.3.5
pytz==2025.2
PyYAML==6.0.3
Quart==0.20.0
quart-cors==0.8.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.2
scipy==1.16.3
selectolax==0.4.0
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart_open==7.5.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
timezonefinder==8.1.0
timezonefinderL==4.0.2
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typer==0.23.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
yarl==1.22.0
yt-dlp==2025.10.22
zipp==3.23.0
